requirements.txt
mysql-connector-python==8.0.33
inputimeout==1.0.4
twilio==8.10.0
opencv-python==4.9.0.80
numpy==1.24.4
Pillow==10.2.0
